Use locks to detect connections which are in an unknown transaction state
Partially Fixes #48164

Introduces the concept of a transaction lock. This is used to detect
cases where `within_new_transaction` raises, potentially leaving the
connection in a transaction, and the attempt to discard the connection
fails.

The idea is to store and acquire a new lock before attempting to
begin a transaction, and to only discard the lock once the
transaction state is definitely known. If `within_new_transaction`
raises, without the lock being discarded, subsequent attempts to
create a transaction will detect this case and discard the connection.

Note that no such check is performed in other methods which may use a
connection in an unknown transaction state. For instance, it will still
be possible to invoke `execute`.