Rewrite Application Templates section [ci-skip]
This modifies the example application template to only use a single gem,
to run generator calls in an `after_bundle` block, and to include a
`db:migrate` call after running the generators.

This also simplifies the section's prose, and mentions
`bin/rails app:template` for running templates in existing apps.