Relocate Command Line Options section [ci-skip]
Adding a command line option is a natural follow-up to creating a
generator, and the example of specifying a command line option builds on
the example from the "Creating Generators with Generators" section.
Therefore, this commit puts the "Adding Command Line Options" section
right after the "Creating Generators with Generators" section.

This commit also tweaks the section to improve clarity.