Refactor error highlight detection
Before this commit, some calls to render were hard-coding error
highlight as "not available".  This was causing some error pages to show
the "you should install error highlight" message even though the right
version of error highlight was installed.

This commit adds a delegate method to the DebugView class so that the
debugging related templates can just ask whether or not error highlight
is available via a method call.  That way we don't need to rely on
passing locals everywhere.  The down side is that this change requires
all "rescue" templates to be rendered within the context of a DebugView
class (but I think that's OK)