Add syntax for CTE materialization hint to ActiveRecord
Following on the Arel work in #48261, this commit adds the ability
to provide a materialization hint to a CTE in ActiveRecord. To
support the hint, a more verbose syntax for CTE values is
introduced:

   Post.with(t: { query: Post.where("comments_count > ?", 0), materialized: true })

The CTE value is described as a hash that contains a required
:query key and an optional :materialized key. The set of available
options could be expanded in the future to accomodate, for example,
postgres's various options for recursive CTEs.