Fix test to keep behavior the same after #48397
This test was failing because the parameter `a` in the accept header
is now being correctly extracted as a mime type.

This test was expecting that Rails would extract it an an empty accept
header, that will only happen now if the `q` parameter that has special
meaning in used.