Add tests around has_one changing the target and validators
Ref: https://github.com/rails/rails/issues/48330

When replacing the has_one target, it seems more correct to first
delete the old record, so that if the associated model has a uniqueness
validator, it won't fail.

Both the delete and the insert are in a single transaction, so if the new
record fail to be saved, the transaction will be rolled back, which seems
correct.

If `dependent: :destroy` isn't set, Active Record will try to orphan the
record, which may or may not be valid depending on the schema and validators.

Co-Authored-By: zzak <zzakscott@gmail.com>