Add `reload_attributes` to Persistence
Add ability for applications to specify a list of attributes that will be reloaded with a SQL `RETURNING` clause when new records are inserted. Long term the plan for databases that do not support RETURNING is to execute an additional SELECT statement.

Also handles a minor edge case if application code attempts to write to a virtual column. The values from `RETURNING` will always be used to update the instance's attributes.