Deprecate `#remove_connection` in favor of `#remove_connection_pool`.
Rails 7.2 will remove support for `#remove_connection` in favor
of `#remove_connection_pool`. This deprecation also represents a
change in behavior. `#remove_connection_pool` will not accept a `name`
argument as it should always be called on the class that owns the
pool an application wants removed. In addition it will no longer reset
a subclass with it's own connection to the parent pool.

Previous behavior:

```ruby
class ActiveRecord::Base
end

class Post < ActiveRecord::Base
  establish_connection :post_db
end

Post.remove_connection
=> would reassign Post to `ActiveRecord::Base`  and `Post.connected?` returns `true`.
```

New behavior:

```ruby
class ActiveRecord::Base
end

class Post < ActiveRecord::Base
  establish_connection :post_db
end

Post.remove_connection
=> Removes the pool keyed on `Post` and `Post.connected?` returns `nil`.
```

If the `Post` class does not have its own connection and `Post.remove_connection`
is called, the behavior is unchanged. Active Record will not remove the pool
belonging to `ActiveRecord::Base`. To remove the parent pool, applications
should call `ActiveRecord::Base.remove_connection` directly.

I chose to add a new method instead of deprecating behavior in the
current one because this behavior change isn't worth the overhead of a
config option. I also think it's more likely to get feedback if this is
a dealbreaker change.

I'm making this change because due to the current design of Active
Record connection management, it does not make sense that
`remove_connection` would reassign connections to the parent class. In a
multi-db application this would be dangerous, we would not want to
assign `AnimalsRecord` to `ActiveRecord::Base` if they are completely
different databases. In addition it does not make sense to take a `name`
argument when the `name` should be derived from the
connection_specification_name on self.