Remove no connection primary key test
This test was added in https://github.com/rails/rails/commit/b3407c86cfccd2cc6258b8879e91600fe25c6e48
at a time when Rails was reading the `primary_keys` off a hash stored on
`connection`. We no longer do that in Active Record so this test isn't
really necessary. I'm removing it because it was flaky. I attempted to
fix it in #48473 but realized today it's not testing the same thing it
was before. After discussing it with Matthew, I've decided to remove it.