Autoload ActiveRecord::FixtureSet
When eager_load is enabled and something (like the frozen_record gem)
tries to load yaml files that use `ActiveRecord::FixtureSet.identify(...)`
we'll get an uninitialized constant error.

Move requires for constants in the ActiveRecord::FixtureSet namespace
to below the creation of the class to avoid circular requires.
See ActiveSupport::Deprecation for a similar instance of this:
https://github.com/rails/rails/blob/7dd67dc285fcdbb788cbdcdcaf53f6ce588bf7e5/activesupport/lib/active_support/deprecation.rb#L35-L54