Wrap active_record_encryption.configuration on an after initialize
This commit changes the active_record_encryption.configuration
to be set on an after initialize block. After PR #44873, we can
configure now hash_digest_class that will be used on
ActiveRecord::Encryption. But the docs said that we could change
it via an initializer by setting
Rails.application.config.active_record.encryption.hash_digest_class.
This wasn't working because the railtie
"active_record_encryption.configuration" was being set before the
initializer was loaded. This changes this behave to configure
the active_record encryption after the initialize was loaded.