Add GlobalID to the API docs
GlobalID is a dependency of Active Job and Action Text. It adds `to_gid`
and `to_sgid` methods to Active Model and Active Record classes.

Trying to find the documentation of these methods can be confusing as
they are not present in the API docs.

By symlinking the globalid gem from the Gemfile we can include it in the
API documentation. After the `rdoc` task has run we can remove the
symlink. I'd prefer to remove to symlink in an `ensure` block
but that doesn't seem to be supported by Rake.