Move 6.1 cache format warning to where its used
A deprecation warning was [added][1] to ensure that applications
manually setting `config.active_support.cache_format_version` to `6.1`
will be aware that they need to migrate. However, if an app is not using
a `config.load_defaults` of `7.0` or greater, this warning will never be
triggered.

This commit moves the deprecation warning to where the `cache_format`
value gets used to cover both cases.

[1]: 2ba3ac29c36f4c6d23b1dd302584f15739ff2aff