Raise an exception on has_one failed to autosave
Fixes #36833
Fixes #48633

Previously, failure to autosave a has_one association returned
nil instead of raising an exception.

    class Supplier < ActiveRecord::Base
      has_one :account, autosave: true
    end

    class Account < ActiveRecord::Base
      before_save { throw(:abort) }
    end

    supplier = Supplier.new
    supplier.build_account
    supplier.save! # => nil

But it's counter-intuitive for save! to fail without raising
an exception.

This commit changes save_has_one_association to raise
ActiveRecord::RecordInvalid instead of ActiveRecord::Rollback.
This allows it to raise an exception if it fails to autosave
the has_one association with save!.

This commit is very similar to [the commit to has_many][1].

[1]: bbcd707aefe5da137137a8deb13908ec9a7db77d

Co-authored-by: Takayuki Nakata <f.seasons017@gmail.com>