Omit list for Cache::Store#fetch "Dynamic Options" [ci-skip]
The extra indentation on this list causes it to be rendered as code
instead of as an unordered list.  But, furthermore, the items in this
list should be setter methods for `ActiveSupport::Cache::WriteOptions`,
not symbols.  Since the `ActiveSupport::Cache::WriteOptions` class is
linked in the preceding paragraph, we can simply omit this list.

(cherry picked from commit 46c42a2ff93c9666580b81512d61ac562f6b9c19)