[7-0-stable] Backport: Remove no connection primary key test
This test was added in rails@b3407c8
at a time when Rails was reading the `primary_keys` off a hash stored on
`connection`. We no longer do that in Active Record so this test isn't
really necessary. I'm removing it because it was flaky. I attempted to
fix it in rails#48473 but realized today it's not testing the same thing it
was before. After discussing it with Matthew, I've decided to remove it.

Co-authored-by: eileencodes <eileencodes@gmail.com>